refactor: Reorganise jupyter related config
* Jupyter server and lab extension points are moved to jupyterlab module

* Names are modified appropriately in config files
