--- jupyter-config/nbconfig/notebook.d/jupytext.json
@@ -1,5 +1,0 @@
-{
-  "load_extensions": {
-    "jupytext/index": true
-  }
-}
