+++ jupyterlab/jupyter-config/jupyter_server_config.d/jupyterlab-jupytext.json
@@ -1,7 +1,7 @@
 {
   "ServerApp": {
     "jpserver_extensions": {
-      "jupytext": true
+      "jupyterlab_jupytext": true
     }
   }
 }
